move stargz code to a separate package
