+++ pkg/stargz/resolver.go
