+++ pkg/stargz/resolver_test.go
