+++ pkg/stargz/testdata/config/nydus.json
