+++ pkg/stargz/testdata/stargz.index.json
