+++ pkg/stargz/testdata/stargzfooter.bin
