+++ pkg/stargz/testdata/stargztoc.bin
